Add shared PixiTomlEditor with UI/TOML mode toggle for workspace editing
Extract the dual-mode pixi.toml editor (UI Mode with packages table + TOML Mode
with raw textarea) into a reusable PixiTomlEditor component. Both the workspace
creation form and workspace detail editing now use this shared component instead
of duplicating UI or showing only a raw textarea.

Also make the Nebi logo clickable to navigate home, and add an Edit button to
the workspace detail header next to Publish/Share.
