+++ frontend/src/components/layout/Layout.tsx
@@ -24,11 +24,13 @@ export const Layout = () => {
         <div className="container mx-auto px-4 py-4">
           <div className="flex items-center justify-between">
             <div className="flex items-center gap-8">
-              <img
-                src="/nebi-logo.png"
-                alt="Nebi"
-                className="h-10 w-auto"
-              />
+              <NavLink to="/workspaces">
+                <img
+                  src="/nebi-logo.png"
+                  alt="Nebi"
+                  className="h-10 w-auto"
+                />
+              </NavLink>
               <nav className="flex gap-1">
                 <NavLink to="/workspaces">
                   {({ isActive }) => (
