+++ frontend/src/pages/Workspaces.tsx
@@ -6,16 +6,11 @@ import { useModeStore } from '@/store/modeStore';
 import { workspacesApi } from '@/api/workspaces';
 import { Button } from '@/components/ui/button';
 import { Input } from '@/components/ui/input';
-import { Textarea } from '@/components/ui/textarea';
 import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card';
 import { Badge } from '@/components/ui/badge';
 import { ConfirmDialog } from '@/components/ui/confirm-dialog';
-import { Loader2, Plus, Trash2, X, Edit, FileCode, Cloud, Monitor, HardDrive } from 'lucide-react';
-
-interface Package {
-  name: string;
-  version: string;
-}
+import { PixiTomlEditor } from '@/components/workspace/PixiTomlEditor';
+import { Loader2, Plus, Trash2, X, Edit, Cloud, Monitor, HardDrive } from 'lucide-react';
 
 type UnifiedWorkspace = {
   id: string;
@@ -49,28 +44,6 @@ platforms = ["osx-arm64", "linux-64"]
 python = ">=3.11"
 `;
 
-const buildPixiToml = (packages: Package[], wsName: string): string => {
-  const dependenciesLines = packages
-    .filter(pkg => pkg.name.trim())
-    .map(pkg => {
-      if (pkg.version.trim()) {
-        return `${pkg.name} = "${pkg.version}"`;
-      }
-      // If no version specified, use "*" (any version)
-      return `${pkg.name} = "*"`;
-    })
-    .join('\n');
-
-  return `[workspace]
-name = "${wsName}"
-channels = ["conda-forge"]
-platforms = ["osx-arm64", "linux-64", "win-64"]
-
-[dependencies]
-${dependenciesLines || 'python = ">=3.11"'}
-`;
-};
-
 export const Workspaces = () => {
   const navigate = useNavigate();
   const { data: workspaces, isLoading } = useWorkspaces();
@@ -88,10 +61,6 @@ export const Workspaces = () => {
   const [newWsName, setNewWsName] = useState('');
   const [localPath, setLocalPath] = useState('');
   const [pixiToml, setPixiToml] = useState(DEFAULT_PIXI_TOML);
-  const [createMode, setCreateMode] = useState<'ui' | 'toml'>('ui');
-  const [packages, setPackages] = useState<Package[]>([{ name: 'python', version: '>=3.11' }]);
-  const [newPackageName, setNewPackageName] = useState('');
-  const [newPackageVersion, setNewPackageVersion] = useState('');
 
   const [showEdit, setShowEdit] = useState(false);
   const [editWsId, setEditWsId] = useState<string | null>(null);
@@ -146,9 +115,7 @@ export const Workspaces = () => {
 
     setError('');
     try {
-      const tomlContent = createMode === 'ui'
-        ? buildPixiToml(packages, newWsName)
-        : pixiToml;
+      const tomlContent = pixiToml;
 
       if (createTarget === 'server' && isRemoteConnected) {
         await createRemoteMutation.mutateAsync({
@@ -169,7 +136,6 @@ export const Workspaces = () => {
       setNewWsName('');
       setLocalPath('');
       setPixiToml(DEFAULT_PIXI_TOML);
-      setPackages([{ name: 'python', version: '>=3.11' }]);
       setShowCreate(false);
 
       if (createTarget === 'local' || !isRemoteConnected) {
@@ -180,24 +146,6 @@ export const Workspaces = () => {
       const errorMessage = error?.response?.data?.error || 'Failed to create workspace. Please try again.';
       setError(errorMessage);
     }
-  };
-
-  const handleAddPackage = () => {
-    if (!newPackageName.trim()) return;
-    setPackages([...packages, { name: newPackageName, version: newPackageVersion }]);
-    setNewPackageName('');
-    setNewPackageVersion('');
-  };
-
-  const handleRemovePackageFromList = (index: number) => {
-    setPackages(packages.filter((_, i) => i !== index));
-  };
-
-  const handleModeSwitch = (mode: 'ui' | 'toml') => {
-    if (mode === 'toml' && createMode === 'ui') {
-      setPixiToml(buildPixiToml(packages, newWsName || 'my-project'));
-    }
-    setCreateMode(mode);
   };
 
   const handleDelete = async () => {
@@ -274,131 +222,6 @@ export const Workspaces = () => {
       </div>
     );
   }
-
-  // Render the package list / TOML form (shared between create targets)
-  const renderCreateForm = () => (
-    <>
-      {/* UI / TOML Mode Toggle */}
-      <div className="flex gap-2 p-1 bg-muted rounded-lg w-fit">
-        <Button
-          type="button"
-          variant={createMode === 'ui' ? 'default' : 'ghost'}
-          size="sm"
-          onClick={() => handleModeSwitch('ui')}
-          className="gap-2"
-        >
-          <Plus className="h-4 w-4" />
-          UI Mode
-        </Button>
-        <Button
-          type="button"
-          variant={createMode === 'toml' ? 'default' : 'ghost'}
-          size="sm"
-          onClick={() => handleModeSwitch('toml')}
-          className="gap-2"
-        >
-          <FileCode className="h-4 w-4" />
-          TOML Mode
-        </Button>
-      </div>
-
-      {createMode === 'ui' ? (
-        <div className="space-y-4">
-          <div className="space-y-2">
-            <label className="text-sm font-medium">Packages</label>
-            <div className="border rounded-lg overflow-hidden">
-              <table className="w-full">
-                <thead className="bg-muted/50 border-b">
-                  <tr>
-                    <th className="text-left p-3 text-sm font-medium">Name</th>
-                    <th className="text-left p-3 text-sm font-medium">Version Constraint</th>
-                    <th className="w-16"></th>
-                  </tr>
-                </thead>
-                <tbody className="divide-y">
-                  {packages.map((pkg, index) => (
-                    <tr key={index} className="hover:bg-muted/30">
-                      <td className="p-3">
-                        <span className="font-mono text-sm">{pkg.name}</span>
-                      </td>
-                      <td className="p-3">
-                        <span className="font-mono text-sm text-muted-foreground">
-                          {pkg.version || '-'}
-                        </span>
-                      </td>
-                      <td className="p-3">
-                        <Button
-                          type="button"
-                          variant="ghost"
-                          size="sm"
-                          onClick={() => handleRemovePackageFromList(index)}
-                        >
-                          <Trash2 className="h-4 w-4" />
-                        </Button>
-                      </td>
-                    </tr>
-                  ))}
-                </tbody>
-              </table>
-            </div>
-          </div>
-
-          {/* Add Package Form */}
-          <div className="flex gap-2">
-            <Input
-              placeholder="Package name (e.g., numpy)"
-              value={newPackageName}
-              onChange={(e) => setNewPackageName(e.target.value)}
-              onKeyDown={(e) => {
-                if (e.key === 'Enter') {
-                  e.preventDefault();
-                  handleAddPackage();
-                }
-              }}
-            />
-            <Input
-              placeholder="Version (e.g., >=1.24.0)"
-              value={newPackageVersion}
-              onChange={(e) => setNewPackageVersion(e.target.value)}
-              onKeyDown={(e) => {
-                if (e.key === 'Enter') {
-                  e.preventDefault();
-                  handleAddPackage();
-                }
-              }}
-              className="w-64"
-            />
-            <Button
-              type="button"
-              onClick={handleAddPackage}
-              disabled={!newPackageName.trim()}
-            >
-              <Plus className="h-4 w-4 mr-2" />
-              Add Package
-            </Button>
-          </div>
-          <p className="text-xs text-muted-foreground">
-            Add packages with optional version constraints (e.g., {'>'}=1.24.0, ~=2.0.0, 3.11.*)
-          </p>
-        </div>
-      ) : (
-        <div className="space-y-2">
-          <label className="text-sm font-medium">pixi.toml Configuration</label>
-          <Textarea
-            placeholder="Enter your pixi.toml content"
-            value={pixiToml}
-            onChange={(e) => setPixiToml(e.target.value)}
-            rows={12}
-            required
-            className="font-mono text-sm"
-          />
-          <p className="text-xs text-muted-foreground">
-            Define your project dependencies and configuration in TOML format
-          </p>
-        </div>
-      )}
-    </>
-  );
 
   return (
     <div className="space-y-6">
@@ -491,7 +314,7 @@ export const Workspaces = () => {
                 </div>
               )}
 
-              {renderCreateForm()}
+              <PixiTomlEditor tomlValue={pixiToml} onTomlChange={setPixiToml} workspaceName={newWsName || 'my-project'} />
 
               <div className="flex gap-2 justify-end">
                 <Button type="button" variant="outline" onClick={() => setShowCreate(false)}>
@@ -541,19 +364,7 @@ export const Workspaces = () => {
                 />
               </div>
 
-              <div className="space-y-2">
-                <label className="text-sm font-medium">pixi.toml Configuration</label>
-                <Textarea
-                  placeholder="Enter your pixi.toml content"
-                  value={editPixiToml}
-                  onChange={(e) => setEditPixiToml(e.target.value)}
-                  rows={12}
-                  required
-                />
-                <p className="text-xs text-muted-foreground">
-                  Editing will delete the old workspace and create a new one with updated configuration
-                </p>
-              </div>
+              <PixiTomlEditor tomlValue={editPixiToml} onTomlChange={setEditPixiToml} workspaceName={editWsName || 'my-project'} />
 
               <div className="flex gap-2 justify-end">
                 <Button type="button" variant="outline" onClick={() => setShowEdit(false)}>
